This prevents the double links in linked-heading
Currently if a heading already contains an anchor, we wrap it in another
anchor. This leads to invalid HTML, and some weird browser behaviour
where both links are attempted to be followed (in Chrome at least)

I think that linked-heading should just ignore the heading element if it
already contains a link. The only potential issue here is whether this
may constitute a breaking change to any of our users, my suspicion is
that it won't - it's technically a bug fix.