Remove most `minmax` uses to work around Chrome 80 bug.
Chrome 80 appears to have introduced a Grid bug which impacts
o-layout:
https://bugs.chromium.org/p/chromium/issues/detail?id=1050307&q=grid&can=2

This PR attempts to work around that bug by removing most uses
of `minmax`. These changes impact the documentation and query
layout.

It appears we can remove these `minmax` uses safely because:
- The sidebars should not change in width according to content. When
a user navigates around the page the main content area should be in
the same position.
- Given the sidebars have a set width, the main content area does
not need a min size -- it can take the remaining space with 1fr.

If content overflows the main grid area `minmax(32.5rem,1fr)` would
allow the content to overflow without increasing the grid. With
a main area of `1fr` it appears the grid increases in width with
overflow content and visibly breaks the layout more dramatically.
Instead don't allow overflow on the main area and trust users
to address overflow as they see fit for their project.