Replace flexible layout columns with a `minmax` function.
`1fr` is the same as `minmax(auto, 1fr)`. It sizes the column
automatically, according the the dimentions of the column content,
and distributes one fraction of any remaining space. This means
a column becomes too large if there is some long content within it.

What we actually want is for our columns to take a fraction of
remaining space, and for any overflow to overflow `minmax(0, 1fr)`.

https://drafts.csswg.org/css-grid/#valdef-grid-template-columns-flex