Allow base/landing layout main CSS grid columns to shrink to `0`.
By setting `minmax(auto, [max-container-size])` we are saying the
main column should be greater than or equal to its smallest size
which doesn’t lead to overflow (auto) and less than or equal to
the max container size (max). For pages which have long content,
such as a code block, this leads to the whole layout scrolling
horizontally on viewports smaller than the max container size --
although the code block could scroll, `auto` tries to accommodate
the element without causing scroll. The behaviour we actually want
is for any overflowing content to overflow, not to increase the
width of the whole layout beyond the viewport. We can do this with:
`minmax(0, [max-container-size])`

https://www.w3.org/TR/css-grid-1/#valdef-grid-template-columns-minmax