Add master brand support for default, bleed, documentation layouts.
This will be used first for the legal pages of ft.com via a
Wordpress theme.

The documentation layout is the only layout required but the default
layout is a given, that the documentation layout builds on, and the
bleed layout is minimal. Other layouts have not been added as it
is likely we would want design input for more complex usecases.