Fix hero image background (currently obscured)
Apply backgorund color to the body rather than `o-layout` to avoid
obscuring the hero element, which makes use of a negative z-index
at the time of writing.

Technically... this breaks with the spec, as does o-normalize, by
applying styles outside its owned scope. This feels reasonable for
a component which should help style whole page layouts but, in
the future, we may way to update o-layout to have a class on the
body element.