feat: improve performance and logic around how to highlight headings in the sidebar
This changes the implementation from using a scroll event listener to using a series of intersection observers, which run off the main thread and are built to do this work in a performant manner.

I also took this time as an opportunity to improve the logic used to determine which heading to highlight.

The logic being used now is:
- Figure out what headings are above the intersection target and what ones are below it and place them in two arrays `above` and `below`.
- Figure out the scrolling direction
- If scrolling down the page, pick the last heading which is in the `above` array
- Else pick the first heading in the `below` array

When the page is scrolled to the bottom, then the last heading is highlighted, always. This is different from the previous logic which for certain document would never highlight the last heading.